Painter: Minor optimizations
git-svn-id: svn://ultimatepp.org/upp/trunk@12462 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Draw/RescaleFilter.cpp
@@ -148,7 +148,7 @@ Image RescaleFilter(const Image& img, int cx, int cy,
 
 static double sNearest(double x)
 {
-	return x >= -0.5 && x <= 0.5;
+	return (double)(x >= -0.5 && x <= 0.5);
 }
